requirements.txt
numpy
opencv-python
pyyaml
ultralytics
onnxruntime
